pyproject.toml
[project]
name = "nutri-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.0.2",
    "openai>=1.107.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
